# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "향수 정보 서비스 A.fume Server Api 문서"
  version: 1.0.0
  title: "Swagger A.fume Server"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "heesung6701@naver.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "{SERVER_URL}:{SERVER_PORT}"
basePath: "/A.fume/api/0.0.1"
...
schemes:
#- "https"
- "http"
# security:
# - APIKeyHeader: []
paths:
...
  /brand:
    get:
      tags:
      - "brand"
      summary: "브랜드 전체 조회"
      description: "브랜드 리스트 반환"
      operationId: "getBrandList"
      produces:
      - "application/json"
      responses:
        "200":
          description: "성공"
          schema:
            type: object
            properties:
              message:
                type: string
                example: 브랜드 조회 성공
              data:
                type: "array"
                items:
                  allOf:
                  - $ref: "#/definitions/BrandInfo"
                  - type: "object"
                    properties:
                      brandIdx:
                        type: "integer"
                    example:
                      brandIdx: 1
        "401":
          description: "API key is missing or invalid"
          headers:
            WWW_Authenticate:
              type: "string"
      x-swagger-router-controller: "Brand"
    post:
      tags:
      - "brand"
      summary: "브랜드 삽입"
      description: "브랜드 삽입"
      operationId: "postBrand"
      security:
        - userToken: []
      x-security-scopes:
        - admin
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Post new brand info"
        required: false
        schema:
          $ref: "#/definitions/BrandInfo"
      responses:
        "200":
          description: "성공"
          schema:
            type: object
            properties:
              message:
                type: string
                example: 브랜드 추가 성공
              data:
                type: integer
                example: 1
                description: 브랜드 brandIdx
        "401":
          description: "API key is missing or invalid"
          headers:
            WWW_Authenticate:
              type: "string"
      deprecated: true
      x-swagger-router-controller: "Brand"
  /brand/{brandIdx}:
    get:
      tags:
      - "brand"
      summary: "브랜드 조회"
      description: "브랜드 조회"
      operationId: "getBrandByIdx"
      security:
        - userToken: []
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "brandIdx"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "성공"
          schema:
            type: object
            properties:
              message:
                type: string
                example: 브랜드 세부 조회 성공
              data:
                allOf:
                - $ref: "#/definitions/BrandInfo"
                - type: "object"
                  properties:
                    brandIdx:
                      type: "integer"
                  example:
                    brandIdx: 1
        "401":
          description: "API key is missing or invalid"
          headers:
            WWW_Authenticate:
              type: "string"
      x-swagger-router-controller: "Brand"
    put:
      tags:
      - "brand"
      summary: "브랜드 수정"
      description: "브랜드 수정"
      operationId: "putBrand"
      security:
        - userToken: []
      x-security-scopes:
        - admin
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "brandIdx"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/BrandInfo"
      responses:
        "200":
          description: "성공"
        "401":
          description: "API key is missing or invalid"
          headers:
            WWW_Authenticate:
              type: "string"
      deprecated: true
      x-swagger-router-controller: "Brand"
    delete:
      tags:
      - "brand"
      summary: "브랜드 삭제"
      description: "브랜드 삭제"
      operationId: "deleteBrand"
      security:
        - userToken: []
      x-security-scopes:
        - admin
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "brandIdx"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "성공"
        "401":
          description: "API key is missing or invalid"
          headers:
            WWW_Authenticate:
              type: "string"
      deprecated: true
      x-swagger-router-controller: "Brand"
  /ingredient:
    get:
      tags:
      - "ingredient"
      summary: "향료 목록 조회"
      description: "향료 목록 반환"
      operationId: "getIngredientList"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "성공"
          schema:
            type: "object"
            properties:
              message:
                type: string
                example: "ingredient 목록 조회 성공"
              data:
                type: "array"
                items:
                  allOf:
                    - $ref: "#/definitions/IngredientInfo"
                    - type: "object"
                      properties:
                        ingredientIdx:
                          type: "integer"
                      example:
                        ingredientIdx: 333
        "401":
          description: "API key is missing or invalid"
          headers:
            WWW_Authenticate:
              type: "string"
      x-swagger-router-controller: "Ingredient"
    post:
      tags:
      - "ingredient"
      summary: "향료 삽입"
      description: "향료 삽입"
      operationId: "postIngredient"
      security:
        - userToken: []
      x-security-scopes:
        - admin
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Post new ingredient info"
        required: false
        schema:
          allOf:
            - $ref: "#/definitions/IngredientInfo"
            - type: "object"
              properties:
                seriesName:
                  type: "string"
              example:
                seriesName : "우디"
      responses:
        "200":
          description: "성공"
          schema:
            type: object
            properties:
              message:
                type: string
                example: ingredient 추가 성공
              data:
                type: integer
                example: 133
                description: ingredient idx
        "401":
          description: "API key is missing or invalid"
          headers:
            WWW_Authenticate:
              type: "string"
      x-swagger-router-controller: "Ingredient"
      deprecated: true
  /ingredient/{ingredientIdx}:
    get:
      tags:
      - "ingredient"
      summary: "특정 향료 조회"
      description: "특정 향료 조회"
      operationId: "getIngredientByIdx"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "ingredientIdx"
        in: "path"
        description: "향료 ID"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "성공"
          schema:
            type: object
            properties:
              message:
                type: string
                example: ingredient 개별 조회 성공
              date :
                allOf:
                - $ref: "#/definitions/IngredientInfo"
                - type: "object"
                  properties:
                    ingredientIdx:
                      type: "integer"
                  example:
                    ingredientIdx: 333
        "401":
          description: "API key is missing or invalid"
          headers:
            WWW_Authenticate:
              type: "string"
      x-swagger-router-controller: "Ingredient"
    put:
      tags:
      - "ingredient"
      summary: "향료 수정"
      description: "향료 수정"
      operationId: "putIngredient"
      security:
        - userToken: []
      x-security-scopes:
        - admin
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "ingredientIdx"
        in: "path"
        description: "향료 ID"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "Updated series info"
        required: false
        schema:
          $ref: "#/definitions/IngredientInfo"
      responses:
        "200":
          description: "성공"
          schema:
            type: object
            properties:
              message: 
                type: string
                example: ingredient 수정 성공
        "401":
          description: "API key is missing or invalid"
          headers:
            WWW_Authenticate:
              type: "string"
      x-swagger-router-controller: "Ingredient"
      deprecated: true
    delete:
      tags:
      - "ingredient"
      summary: "향료 삭제"
      description: "향료 삭제"
      operationId: "deleteIngredient"
      security:
        - userToken: []
      x-security-scopes:
        - admin
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "ingredientIdx"
        in: "path"
        description: "향료 ID"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "성공"
          schema:
            type: object
            properties:
              message:
                type: string
                example: ingredient 삭제 성공
        "401":
          description: "API key is missing or invalid"
          headers:
            WWW_Authenticate:
              type: "string"
      x-swagger-router-controller: "Ingredient"
      deprecated: true
...
securityDefinitions:
  userToken:
      type: apiKey
      name: x-access-token
      in: header
      description: /Bearer {token}/ 를 입력해주시면 됩니다.
  APIKeyHeader:
    type: apiKey
    name: X-API-Key
    in: header
definitions:
...
  IngredientInfo:
    type: "object"
    properties:
      name:
        type: "string"
      englishName:
        type: "string"
      description:
        type: "string"
    example:
      name: "씨쏠트"
      englishName: "Sea Salt"
      description: "바다 소금"
...
  BrandInfo:
    type: "object"
    properties:
      name:
        type: "string"
      englishName:
        type: "string"
      startCharacter:
        type: "string"
      description:
        type: "string"
      imageUrl:
        type: "string"
    example:
      name: "(테스트)조말론"
      englishName: "(테스트)Jo Malone"
      startCharacter: "ㅌ"
      description: "(테스트)런던 브랜드"
      imageUrl: "(테스트)"
...